refractor med plan functions
